# src/main/java/com/sevenkingdoms/model/enums/ActionType.java
package com.sevenkingdoms.model.enums;

public enum ActionType {
    NORMALE,    // Azione base del personaggio
    POTENZIATA, // Azione potenziata (costa Favore Papale)
    PASSIVA     // Azione passiva (risolta nella fase 2)
}